refactor: edit the code of Invoice.

+++ AquaEngine.API/Sales/Domain/Model/Queries/GetInvoiceByIdQuery.cs
@@ -1,3 +1,3 @@
 namespace AquaEngine.API.Sales.Domain.Model.Queries;
 
-public record GetInvoiceByIdQuery(int Id);
+public record GetInvoiceByIdQuery(long Id);
